requirements.txt
streamlit
pillow
inference-sdk
opencv-python-headless